ENH: Rename repo to 'xray-vision'
