+++ examples/xrf_gui/file_gui.py
@@ -2,12 +2,12 @@ __author__ = 'edill'
 
 import enaml
 from enaml.qt.qt_application import QtApplication
-from bubblegum.xrf.model.xrf_model import XRF
+from xray_vision.xrf.model.xrf_model import XRF
 
 def run():
     app = QtApplication()
     with enaml.imports():
-        from bubblegum.xrf.view.file_view import FileGui
+        from xray_vision.xrf.view.file_view import FileGui
 
     view = FileGui()
     view.xrf_model1 = XRF()
