+++ examples/xrf_gui/xrf_gui.py
@@ -2,12 +2,12 @@ __author__ = 'edill'
 
 import enaml
 from enaml.qt.qt_application import QtApplication
-from bubblegum.xrf.model.xrf_model import XRF
+from xray_vision.xrf.model.xrf_model import XRF
 
 def run():
     app = QtApplication()
     with enaml.imports():
-        from bubblegum.xrf.view.xrf_view import XrfGui
+        from xray_vision.xrf.view.xrf_view import XrfGui
 
     view = XrfGui()
     view.xrf_model = XRF()
