+++ xray_vision/backend/mpl/stack_1d.py
