+++ xray_vision/qt_widgets/control_widgets.py
