+++ xray_vision/qt_widgets/displaydict.py
