+++ xray_vision/qt_widgets/real_time.py
