+++ xray_vision/qt_widgets/util.py
